pyproject.toml
[project]
name = "pokemon-coloring-page"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27.2",
    "pillow>=10.4.0",
    "prompt-toolkit>=3.0.47",
]
